backend/requirements.txt
flask
flask-cors
werkzeug
mysql-connector-python
